refactor(foot): StAllCalc未抽出8セクションを分割（Phase 2b）
Phase 2（ここからバナー約46セクション）の調査スコープが後半の「％補正」
パートに限定されていたため、前半の基礎値算出パート（ATK/MaxHP/MaxSP/
DEF/MDEF/HIT算出、計700行）が見落とされていた。同じ手法で
foot-stallcalc-{atk-base,maxhp-maxsp,def,mdef-hit}.js へ分割。

副次: n_A_WeaponZokusei（UpdateEquipItemDataByHtml書込・ATK算出読取）を
roro-state.js へ移設。head.js への既存の直接import(ApplyElementRatio)は
新ファイルではhead-bridge.js経由に変更（既存箇所はスコープ外で維持）。
