+++ roro/m/js/foot.js
@@ -929,6 +929,11 @@ import { ApplyGuaranteedHitUpPercent } from './foot-stallcalc-guaranteed-hit.js'
 import {
     ApplyExpUpPercent, ApplyMeleePhysicalDamageReflect, ApplyWeaponAtkUpPercent, ApplySize100Percent,
 } from './foot-stallcalc-exp-reflect-atk-size.js';
+// StAllCalc セクション分割（.claude/context/remaining-work.md「残作業 1」Phase 2b）
+import { ApplyPlayerAtkBase } from './foot-stallcalc-atk-base.js';
+import { ApplyPlayerMaxHpMaxSp } from './foot-stallcalc-maxhp-maxsp.js';
+import { ApplyPlayerDef } from './foot-stallcalc-def.js';
+import { ApplyPlayerMdefHit } from './foot-stallcalc-mdef-hit.js';
 // C-6: JOB 定数（mig.job.h.js の export const 化に伴う）
 import {
          JOB_SERIES_ID_NOVICE, JOB_SERIES_ID_SWORDMAN, JOB_SERIES_ID_MAGICIAN, JOB_SERIES_ID_ARCHER,
@@ -1003,6 +1008,7 @@ import {
          g_refinedArray, set_g_refinedArray, g_objMobConfInput, set_g_objMobConfInput,
          SpeedPotName,
          n_A_JOB, n_A_MATK, set_n_A_MATK, BK_n_A_MATK, set_BK_n_A_MATK, set_g_lucky_over, set_n_CastCutForDisp, n_A_Weapon2Type,
+         set_n_A_WeaponZokusei,
 } from './roro-state.js';
 
 import { n_A_Equip, n_A_card, set_n_A_Equip, set_n_A_card } from './roro-state.js';
@@ -1022,7 +1028,7 @@ import { n_A_Equip, n_A_card, set_n_A_Equip, set_n_A_card } from './roro-state.j
 let SU_POW = 0, SU_STA = 0, SU_WIS = 0, SU_SPL = 0, SU_CON = 0, SU_CRT = 0;
 let n_A_POW = 0, n_A_CON = 0;
 let n_A_SpeedPOT = 0;
-let n_A_WeaponZokusei = 0;
+// n_A_WeaponZokusei は Phase 2b で roro-state.js へ移設（foot-stallcalc-atk-base.js と共有するため）
 let n_B_HIT = 0, n_B_FLEE = 0;
 // g_bSuperNoviceFullWeapon は equip.js が所有（import 済みのライブバインディング）。
 // 書き込みは SetSuperNoviceFullWeapon 経由で行う。
@@ -1311,7 +1317,7 @@ export function UpdateEquipItemDataByHtml() {
 	var calcForm = document.calcForm;
 
 	set_n_A_WeaponType(eval(calcForm.A_WeaponType.value));
-	n_A_WeaponZokusei = eval(calcForm.A_Weapon_zokusei.value);
+	set_n_A_WeaponZokusei(eval(calcForm.A_Weapon_zokusei.value));
 	set_n_A_Arrow(parseInt(HtmlGetObjectValueById("OBJID_SELECT_ARROW", ARROW_ID_NONE)));
 
 	// 装備部位配列初期化
@@ -1948,706 +1954,10 @@ export function StAllCalc(){
 
 	StPlusCalc();
 
-//================================================================================================
-// ＡＴＫの算出
-//================================================================================================
-
-	//----------------------------------------------------------------
-	// ステータスＡＴＫの算出
-	//----------------------------------------------------------------
-	var statusAtk = 0;
-	var statusAtkGX = 0;
-
-	switch (n_A_WeaponType) {
-
-	// ＤＥＸ依存武器
-	case ITEM_KIND_BOW:
-	case ITEM_KIND_MUSICAL:
-	case ITEM_KIND_WHIP:
-	case ITEM_KIND_HANDGUN:
-	case ITEM_KIND_RIFLE:
-	case ITEM_KIND_SHOTGUN:
-	case ITEM_KIND_GATLINGGUN:
-	case ITEM_KIND_GRENADEGUN:
-		statusAtk = n_A_DEX + n_A_STR / 5 + n_A_LUK / 3;
-		break;
-
-	// ＳＴＲ依存武器
-	default:
-		statusAtk = n_A_STR + n_A_DEX / 5 + n_A_LUK / 3;
-		break;
-	}
-
-	statusAtk = Math.floor(statusAtk);
-
-	// ＧＸ反動ダメージ用ＡＴＫ
-	statusAtkGX = statusAtk;
-	statusAtkGX += ROUNDDOWN(statusAtkGX * zokusei[n_A_BodyZokusei * 10 +1][0] / 100);
-
-	// 計算した結果をキャラクターデータに保存
-	charaData[CHARA_DATA_INDEX_STATUS_ATK] = statusAtk;
-	charaData[CHARA_DATA_INDEX_STATUS_ATK_GX] = statusAtkGX;
-
-	//----------------------------------------------------------------
-	// 精錬ＡＴＫの算出
-	//----------------------------------------------------------------
-	var refineAtk = 0;
-
-	// 精錬による効果
-	refineAtk += n_A_WeaponLV_seirenATK;
-
-	// 「三次職支援　ストライキング」による、効果
-	if (sklLv = g_confDataSanzi[CCharaConfSanzi.CONF_ID_STRIKING]) {
-		refineAtk += (8 + 2 * sklLv) * n_A_WeaponLV + 5 * g_confDataSanzi[CCharaConfSanzi.CONF_ID_STRIKINGYO_FUYOSKILL_LEVEL_GOKEI];
-	}
-
-	// 「修羅　点穴 -反-」による、効果
-	var valStr = 0;
-	var valDex = 0;
-	if (sklLv = UsedSkillSearch(SKILL_ID_TENKETSU_HAN)) {
-
-		// 持ち替え支援の場合、支援欄のステータスを使用
-		if(n_A_PassSkill4[11] == 2) {
-			valStr = n_A_PassSkill4[30];
-			valDex = n_A_PassSkill4[34];
-		}
-
-		// 持ち替えなし場合、本人のステータスを使用
-		else {
-			valStr = n_A_STR;
-			valDex = n_A_DEX;
-		}
-
-		refineAtk += Math.floor((Math.floor(valStr / 2) + Math.floor(valDex / 4)) * sklLv / 5);
-	}
-
-	// 計算した結果をキャラクターデータに保存
-	charaData[CHARA_DATA_INDEX_REFINE_ATK] = refineAtk;
-
-	//----------------------------------------------------------------
-	// 装備ＡＴＫの算出
-	//----------------------------------------------------------------
-
-	// アイテム特性
-	n_tok[ITEM_SP_ATK_PLUS] += n_tok[244];
-
-	// 装備追加効果、支援効果
-	n_tok[ITEM_SP_ATK_PLUS] += GetStatusModifyAtkPlus();
-
-	// 計算した結果をキャラクターデータに保存
-	charaData[CHARA_DATA_INDEX_WEAPON_ATK] = n_A_Weapon_ATK;
-
-	// 拡張表示用にデータを保存
-	CExtraInfoAreaComponentManager.dispDataMap.set(ITEM_SP_ATK_PLUS, n_tok[ITEM_SP_ATK_PLUS]);
-
-	//----------------------------------------------------------------
-	// 左手ＡＴＫの算出
-	//----------------------------------------------------------------
-
-	var leftAtk = 0;
-	if (n_Nitou) {
-		leftAtk = n_A_Weapon2_ATK + n_A_Weapon2LV_seirenATK;
-	}
-
-	// 計算した結果をキャラクターデータに保存
-	charaData[CHARA_DATA_INDEX_LEFT_ATK] = leftAtk;
-
-	//----------------------------------------------------------------
-	// ステータス欄に表示されない効果の適用
-	//----------------------------------------------------------------
-
-	// 暖かい風
-	if (n_A_PassSkill8[19] == 0) {
-		if (GetLowerJobSeriesID(n_A_JOB) == JOB_ID_TAEGWON) {
-			statusAtk = ApplyElementRatio(mobData, statusAtk, n_A_WeaponZokusei);
-		}
-		else {
-			statusAtk = ApplyElementRatio(mobData, statusAtk, 0);
-		}
-	}
-	else {
-		if (GetLowerJobSeriesID(n_A_JOB) != JOB_ID_TAEGWON) {
-			statusAtk = ApplyElementRatio(mobData, statusAtk, n_A_WeaponZokusei);
-		}
-		else {
-			statusAtk = ApplyElementRatio(mobData, statusAtk, 0);
-		}
-	}
-
-	// インベナム使用時のＡＴＫ補正
-	if (n_A_ActiveSkill == SKILL_ID_ENVENOM) {
-		n_tok[ITEM_SP_ATK_PLUS] += 15 * n_A_ActiveSkillLV;
-	}
-
-	// ポイズンリアクト（反撃）使用時のＡＴＫ補正
-	if ((n_A_ActiveSkill == SKILL_ID_POISON_REACT)) {
-		if ((51 <= mobData[18]) && (mobData[18] < 60)) {
-			n_tok[ITEM_SP_ATK_PLUS] += 75;
-		}
-	}
-
-//================================================================================================
-// ＭａｘＨＰ算出
-//================================================================================================
-
-	//----------------------------------------------------------------
-	// ＨＰの基礎値を取得
-	//----------------------------------------------------------------
-	var maxHp = 0;
-	maxHp = GetHPBase(n_A_JOB, n_A_BaseLV, n_A_PassSkill8[13]);
-
-	// テコンランカーの３倍補正（拳聖以上は効果なし）
-	if ((n_A_JOB == JOB_ID_TAEGWON) && (n_A_BaseLV >= 90)) {
-		if (UsedSkillSearch(SKILL_ID_TAEGWON_RANKER)) {
-			maxHp = maxHp * 3;
-		}
-	}
-
-	// インスピレーション計算用のバックアップを作成しておく
-	var RG_HP_BackUP = 0;
-	RG_HP_BackUP = maxHp;
-
-	//----------------------------------------------------------------
-	// ＶＩＴによる補正
-	//----------------------------------------------------------------
-
-	// インスピレーションのＶＩＴ補正は、基礎ＨＰ増加量に影響を与えない
-	var vitForCalc = n_A_VIT;
-	if (UsedSkillSearch(SKILL_ID_INSPIRATION)
-		|| TimeItemNumSearch(TIME_ITEM_ID_ZETSUBONO_KAMI_MOROCC_CARD)
-		|| TimeItemNumSearch(TIME_ITEM_ID_DEMI_FREYA)
-		|| TimeItemNumSearch(TIME_ITEM_ID_MAKENSHI_SAKRAY_CARD)
-		) {
-		vitForCalc -= ROUNDDOWN(n_A_BaseLV / 10 + n_A_JobLV / 5);
-	}
-	maxHp += ROUNDDOWN(maxHp * vitForCalc / 100);
-
-	
-	//----------------------------------------------------------------
-	// 装備、支援等による補正（＋○○）
-	//----------------------------------------------------------------
-	var maxHpPlus = 0;
-
-	// アイテム特性
-	maxHpPlus += n_tok[ITEM_SP_MAXHP_PLUS];
-	maxHpPlus += n_tok[ITEM_SP_MAXHP_PLUS_GVGTE];
-
-	// 装備によるＶＩＴ上昇分は、ＨＰも同量増加させる
-	maxHpPlus += GetEquippedTotalSPEquip(ITEM_SP_VIT_PLUS);
-	maxHpPlus += GetEquippedTotalSPEquip(ITEM_SP_ALLSTATUS_PLUS);
-
-
-	// 装備追加効果、支援効果
-	maxHpPlus += GetStatusModifyMaxHpPlus();
-
-	// 拡張表示用にデータを保存
-	CExtraInfoAreaComponentManager.dispDataMap.set(ITEM_SP_MAXHP_PLUS, maxHpPlus);
-
-	// 最終的な効果を適用
-	maxHp += maxHpPlus;
-	if (maxHp < 1) maxHp = 1;
-
-
-	//----------------------------------------------------------------
-	// 装備、支援等による補正（＋％）
-	//----------------------------------------------------------------
-	var maxHpPerUp = 0;
-
-	// アイテム特性
-	maxHpPerUp += n_tok[ITEM_SP_MAXHP_UP];
-
-	// 装備追加効果、支援効果
-	maxHpPerUp += GetStatusModifyMaxHpUp();
-
-	// 拡張表示用にデータを保存
-	CExtraInfoAreaComponentManager.dispDataMap.set(ITEM_SP_MAXHP_UP, maxHpPerUp);
-
-	// 最終的な効果を適用
-	maxHp += maxHp * maxHpPerUp / 100;
-
-
-	//----------------------------------------------------------------
-	// ロイヤルガード　インスピレーション使用時の特殊計算
-	//----------------------------------------------------------------
-	if (UsedSkillSearch(SKILL_ID_INSPIRATION)) {
-
-		// インスピレーションの効果が無い状態で、インスピレーションによる上昇量を算出
-		// ここの計算で maxHp の値を使うので、一度に計算できない
-		vartmp = ROUNDDOWN(UsedSkillSearch(SKILL_ID_INSPIRATION) * (600 + maxHp / 20));
-
-		// バックアップしておいたＨＰを元に、インスピレーションの効果込みでＨＰを再計算
-		maxHp = RG_HP_BackUP;
-		maxHp += ROUNDDOWN(maxHp * n_A_VIT / 100);
-		maxHp += maxHpPlus + vartmp;		// ここで、インスピレーションのＨＰ上昇を適用
-		maxHp += ROUNDDOWN(maxHp * maxHpPerUp / 100);
-	}
-	else if (
-		TimeItemNumSearch(TIME_ITEM_ID_ZETSUBONO_KAMI_MOROCC_CARD)
-		|| TimeItemNumSearch(TIME_ITEM_ID_DEMI_FREYA)
-		|| TimeItemNumSearch(TIME_ITEM_ID_MAKENSHI_SAKRAY_CARD)
-		) {
-
-		// インスピレーションの効果が無い状態で、インスピレーションによる上昇量を算出
-		// ここの計算で maxHp の値を使うので、一度に計算できない
-		vartmp = ROUNDDOWN(1 * (600 + maxHp / 20));
-
-		// バックアップしておいたＨＰを元に、インスピレーションの効果込みでＨＰを再計算
-		maxHp = RG_HP_BackUP;
-		maxHp += ROUNDDOWN(maxHp * n_A_VIT / 100);
-		maxHp += maxHpPlus + vartmp;		// ここで、インスピレーションのＨＰ上昇を適用
-		maxHp += ROUNDDOWN(maxHp * maxHpPerUp / 100);
-	}
-
-
-	//----------------------------------------------------------------
-	// 特殊補正
-	//----------------------------------------------------------------
-
-	// バーサーク
-	if (UsedSkillSearch(SKILL_ID_BERSERK)) maxHp *= 3;
-
-
-	//----------------------------------------------------------------
-	// 小数点以下切り落とし
-	//----------------------------------------------------------------
-	maxHp = Math.floor(maxHp);
-
-
-	//----------------------------------------------------------------
-	// 計算した結果をキャラクターデータに保存
-	//----------------------------------------------------------------
-	charaData[CHARA_DATA_INDEX_MAXHP] = maxHp;
-
-//================================================================================================
-// ＭａｘＳＰ算出
-//================================================================================================
-
-	//----------------------------------------------------------------
-	// ＳＰの基礎値を取得
-	//----------------------------------------------------------------
-	var maxSp = 0;
-
-	maxSp = GetSPBase(n_A_JOB, n_A_BaseLV, n_A_PassSkill8[13]);
-
-	// テコンランカーの３倍補正（拳聖以上は効果なし）
-	if ((n_A_JOB == JOB_ID_TAEGWON) && (n_A_BaseLV >= 90)) {
-		if (UsedSkillSearch(SKILL_ID_TAEGWON_RANKER)) {
-			maxSp = maxSp * 3;
-		}
-	}
-
-	//----------------------------------------------------------------
-	// ＩＮＴによる補正
-	//----------------------------------------------------------------
-	maxSp += ROUNDDOWN(maxSp * n_A_INT / 100);
-
-	//----------------------------------------------------------------
-	// 装備、支援等による補正（＋○○）
-	//----------------------------------------------------------------
-	var maxSpPlus = 0;
-
-	// アイテム特性
-	maxSpPlus += n_tok[ITEM_SP_MAXSP_PLUS];
-	maxSpPlus += n_tok[ITEM_SP_MAXSP_PLUS_GVGTE];
-
-	// 装備によるＩＮＴ上昇分は、ＳＰも同量増加させる
-	maxSpPlus += GetEquippedTotalSPEquip(ITEM_SP_INT_PLUS);
-	maxSpPlus += GetEquippedTotalSPEquip(ITEM_SP_ALLSTATUS_PLUS);
-
-
-	// 装備追加効果、支援効果
-	maxSpPlus += GetStatusModifyMaxSpPlus();
-
-	// 拡張表示用にデータを保存
-	CExtraInfoAreaComponentManager.dispDataMap.set(ITEM_SP_MAXSP_PLUS, maxSpPlus);
-
-	// 最終的な効果を適用
-	maxSp += maxSpPlus;
-	if(maxSp <0) maxSp = 0;
-
-	//----------------------------------------------------------------
-	// 装備、支援等による補正（＋％）
-	//----------------------------------------------------------------
-	var maxSpPerUp = 0;
-
-	// アイテム特性
-	maxSpPerUp += n_tok[ITEM_SP_MAXSP_UP];
-
-	// 装備追加効果、支援効果
-	maxSpPerUp += GetStatusModifyMaxSpUp();
-
-	// 拡張表示用にデータを保存
-	CExtraInfoAreaComponentManager.dispDataMap.set(ITEM_SP_MAXSP_UP, maxSpPerUp);
-
-	// 最終的な効果を適用
-	maxSp += Math.floor(maxSp * maxSpPerUp / 100);
-
-	//----------------------------------------------------------------
-	// 計算した結果をキャラクターデータに保存
-	//----------------------------------------------------------------
-	charaData[CHARA_DATA_INDEX_MAXSP] = maxSp;
-
-//================================================================================================
-// 除算Ｄｅｆ算出
-//================================================================================================
-
-	//----------------------------------------------------------------
-	// 基本除算Ｄｅｆ
-	//----------------------------------------------------------------
-	var armorDef = 0;
-
-	// 防具の基本Ｄｅｆ
-
-	// 従来の処理
-	for (idx = EQUIP_REGION_ID_HEAD_TOP; idx <= EQUIP_REGION_ID_ACCESSORY_2; idx++) {
-		armorDef += ItemObjNew[n_A_Equip[idx]][ITEM_DATA_INDEX_POWER];
-	}
-
-	//----------------------------------------------------------------
-	// 装備、支援等による補正（＋○○）
-	//----------------------------------------------------------------
-	var defDivPlus = 0;
-
-	// アイテム特性
-	defDivPlus += n_tok[ITEM_SP_DEF_PLUS];
-
-	// 装備追加効果、支援効果
-	defDivPlus += GetStatusModifyDefDivPlus();
-
-	// 拡張表示用にデータを保存
-	CExtraInfoAreaComponentManager.dispDataMap.set(ITEM_SP_DEF_PLUS, defDivPlus);
-
-
-	//----------------------------------------------------------------
-	// 防具の精錬効果
-	//----------------------------------------------------------------
-	var refineDefTable = [0, 1, 2, 3, 4, 6, 8, 10, 12, 15, 18];
-	var defRefined = 0;
-	defRefined += refineDefTable[n_A_HEAD_DEF_PLUS];
-	defRefined += refineDefTable[n_A_BODY_DEF_PLUS];
-	defRefined += refineDefTable[n_A_SHIELD_DEF_PLUS];
-	defRefined += refineDefTable[n_A_SHOULDER_DEF_PLUS];
-	defRefined += refineDefTable[n_A_SHOES_DEF_PLUS];
-
-
-	//----------------------------------------------------------------
-	// 除算Ｄｅｆの算出
-	//----------------------------------------------------------------
-	var defDiv = 0;
-	defDiv = armorDef + defDivPlus + defRefined;
-
-
-	//----------------------------------------------------------------
-	// 特殊効果の適用（各々乗算扱い）
-	//----------------------------------------------------------------
-
-	// アイテム特性　キャラの防御力１／ｎ
-	if (n_tok[ITEM_SP_DEF_DIVIDE_PENARTY]) {
-		defDiv = Math.floor(defDiv / n_tok[ITEM_SP_DEF_DIVIDE_PENARTY]);
-	}
-
-	// ロードナイト　コンセントレイション
-	if (sklLv = UsedSkillSearch(SKILL_ID_CONCENTRATION)) {
-		defDiv = Math.floor(defDiv * (1 - 0.05 * sklLv));
-	}
-	// 「二次職支援　コンセントレイション」の、効果
-	else if ((sklLv = g_confDataNizi[CCharaConfNizi.CONF_ID_CONCENTRATION]) > 0) {
-		defDiv = Math.floor(defDiv * (1 - 0.05 * sklLv));
-	}
-
-	// ロードナイト　バーサーク
-	if (UsedSkillSearch(SKILL_ID_BERSERK)) {
-		defDiv = 0;
-	}
-
-	/* プレイヤー状態異常設定「永遠の混沌」の効果 */
-	if (g_confDataDebuff[CCharaConfDebuff.CONF_ID_ETERNALCHAOS] > 0) {
-		defDiv = 0;
-	}
-
-	// 囲まれ補正
-	if (n_A_PassSkill8[12] >= 3) {
-		defDiv -= Math.floor(defDiv * (n_A_PassSkill8[12] - 2) * 5 / 100);
-	}
-
-	//----------------------------------------------------------------
-	// 装備、支援等による補正（＋％）
-	//----------------------------------------------------------------
-	var defDivUp = 0;
-
-	// アイテム特性
-	defDivUp += n_tok[ITEM_SP_DEF_UP];
-
-	// 装備追加効果、支援効果
-	defDivUp += GetStatusModifyDefDivUp();
-
-	// 拡張表示用にデータを保存
-	CExtraInfoAreaComponentManager.dispDataMap.set(ITEM_SP_DEF_UP, defDivUp);
-
-	// 最終的な効果を適用
-	defDiv += ROUNDDOWN(defDiv * defDivUp / 100);
-
-	//----------------------------------------------------------------
-	// 計算した結果をキャラクターデータに保存
-	//----------------------------------------------------------------
-	charaData[CHARA_DATA_INDEX_DEF_DIV] = defDiv;
-
-//================================================================================================
-// 減算Ｄｅｆ算出
-//================================================================================================
-
-	//----------------------------------------------------------------
-	// 基本減算Ｄｅｆ
-	//----------------------------------------------------------------
-
-	// プロボック系による倍率補正
-	var rateByProvoke = 100;
-	// オートバーサーク
-	if (UsedSkillSearch(SKILL_ID_AUTO_BERSERK)) {
-		rateByProvoke = 45;
-	}
-	// 「一次職支援　支援プロボック」
-	else if (g_confDataIchizi[CCharaConfIchizi.CONF_ID_SHIEN_PROVOKE]){
-		rateByProvoke = 95 - 5 * g_confDataIchizi[CCharaConfIchizi.CONF_ID_SHIEN_PROVOKE];
-	}
-
-	// 基本値を算出
-	var defMinus = 0;
-	defMinus = Math.floor(((n_A_VIT / 2) * rateByProvoke / 100) + n_A_BaseLV / 2 + n_A_AGI / 5);
-
-	//----------------------------------------------------------------
-	// 特殊効果の適用
-	//----------------------------------------------------------------
-
-	// 「ソーサラー　精霊補助スキル　ソリッドスキン」の、効果
-	if (UsedSkillSearch(SKILL_ID_SERE_SUPPORT_SKILL) == 29){
-		defMinus += defMinus;
-	}
-
-	// 「修羅　点穴 -活-」の、効果
-	if ((sklLv = UsedSkillSearch(SKILL_ID_TENKETSU_KATSU)) > 0) {
-		var valVit = 0;
-
-		// 持ち替え支援の場合、支援欄のステータスを使用
-		if(n_A_PassSkill4[11] == 3) {
-			valVit = n_A_PassSkill4[32];
-		}
-
-		// 持ち替えなしの場合、本人のステータスを使用
-		else {
-			valVit = n_A_VIT;
-		}
-
-		defMinus += Math.floor(valVit / 4) * sklLv;
-	}
-
-	// 「一次職支援　エンジェラス」の、効果
-	if (g_confDataIchizi[CCharaConfIchizi.CONF_ID_ANGELUS]) {
-		defMinus += Math.floor(ROUNDDOWN(n_A_VIT / 2) * (0.05 * g_confDataIchizi[CCharaConfIchizi.CONF_ID_ANGELUS]));
-	}
-
-	// 「ミスリルマジックマント　時限効果」の、効果（ペナルティ）
-	if (TimeItemNumSearch(33)) {
-		defMinus -= Math.floor(defMinus * 20 / 100);
-	}
-
-	// 「ロードナイト　バーサーク」の、効果（ペナルティ）
-	if (UsedSkillSearch(SKILL_ID_BERSERK)){
-		defMinus = 0;
-	}
-
-	/* プレイヤー状態異常設定「永遠の混沌」の効果 */
-	if (g_confDataDebuff[CCharaConfDebuff.CONF_ID_ETERNALCHAOS] > 0) {
-		defMinus = 0;
-	}
-
-	// 囲まれ補正
-	if (n_A_PassSkill8[12] >= 3) {
-		defMinus -= Math.floor(defMinus * (n_A_PassSkill8[12] - 2) * 5 / 100);
-	}
-
-	// 「ジェネティック　Ｓホム　ペインキラー」の、効果
-	if ((bufLv = ExBuffNumSearch(EXBUF_ID_HOM_S_PAINKILLER)) > 0) {
-		var homLv = ExBuffNumSearch(EXBUF_ID_HOM_S_PAINKILLER_HOM_LEVEL);
-
-		defMinus += ROUNDDOWN(100 * bufLv * (132 + homLv) / 100 * n_A_BaseLV / 150);
-	}
-
-	//----------------------------------------------------------------
-	// 計算した結果をキャラクターデータに保存
-	//----------------------------------------------------------------
-	charaData[CHARA_DATA_INDEX_DEF_MINUS] = defMinus;
-
-/**
- * ==================================================================
- * 除算Mdef算出
- * ================================================================== 
- */
-
-	var mdefDiv = 0;
-	var mdefDivIgnoreBuff = 0;
-
-	//----------------------------------------------------------------
-	// 装備、支援等による補正（＋○○）
-	//----------------------------------------------------------------
-	var mdefDivPlus = 0;
-	var mdefDivPlusIgnoreBuff = 0;
-
-	// アイテム特性
-	mdefDivPlus = n_tok[ITEM_SP_MDEF_PLUS];
-	mdefDivPlusIgnoreBuff = mdefDivPlus;
-
-	// 装備追加効果、支援効果
-	mdefDivPlus += GetStatusModifyMdefDivPlus(false);
-	mdefDivPlusIgnoreBuff += GetStatusModifyMdefDivPlus(true);
-
-	// 拡張表示用にデータを保存
-	CExtraInfoAreaComponentManager.dispDataMap.set(ITEM_SP_MDEF_PLUS, mdefDivPlus);
-
-	// 最終的な効果を適用
-	mdefDiv = mdefDivPlus;
-	mdefDivIgnoreBuff = mdefDivPlusIgnoreBuff;
-
-	//----------------------------------------------------------------
-	// 特殊効果の適用（各々乗算扱い）
-	//----------------------------------------------------------------
-
-	// 「修羅　点穴 -反-」の、効果（ペナルティ）
-	if ((sklLv = UsedSkillSearch(SKILL_ID_TENKETSU_HAN)) > 0) {
-		var valInt = 0;
-
-		// 持ち替え支援の場合、支援欄のステータスを使用
-		if(n_A_PassSkill4[11] == 2) {
-			valInt = n_A_PassSkill4[32];
-		}
-
-		// 持ち替えなしの場合、本人のステータスを使用
-		else {
-			valInt = n_A_VIT;
-		}
-
-		if (valInt < 1) valInt = 1;
-
-		// このペナルティによって、Ｍｄｅｆがマイナスになることはない（らしい）
-		vartmp = 1 * Math.floor(200 / valInt) * sklLv;
-		mdefDiv -= Math.min(mdefDiv, vartmp);
-	}
-
-	// 「ロードナイト　バーサーク」の、効果（ペナルティ）
-	if (UsedSkillSearch(SKILL_ID_BERSERK)){
-		mdefDiv = 0;
-	}
-
-	//----------------------------------------------------------------
-	// 装備、支援等による補正（＋％）
-	//----------------------------------------------------------------
-	var mdefDivUp = 0;
-	var mdefDivUpIgnoreBuff = 0;
-
-	// アイテム特性
-	mdefDivUp += n_tok[ITEM_SP_MDEF_UP];
-	mdefDivUpIgnoreBuff += n_tok[ITEM_SP_MDEF_UP];
-
-	// 装備追加効果、支援効果
-	mdefDivUp += GetStatusModifyMdefDivUp(false);
-	mdefDivUpIgnoreBuff += GetStatusModifyMdefDivUp(true);
-
-	// 拡張表示用にデータを保存
-	CExtraInfoAreaComponentManager.dispDataMap.set(ITEM_SP_MDEF_UP, mdefDivUp);
-
-	// 最終的な効果を適用
-	mdefDiv += Math.floor(mdefDiv * mdefDivUp / 100);
-	mdefDivIgnoreBuff += Math.floor(mdefDivIgnoreBuff * mdefDivUpIgnoreBuff / 100);
-
-	//----------------------------------------------------------------
-	// 計算した結果をキャラクターデータに保存
-	//----------------------------------------------------------------
-	charaData[CHARA_DATA_INDEX_MDEF_DIV] = mdefDiv;
-	charaData[CHARA_DATA_INDEX_MDEF_DIV_IGNORE_BUFF] = mdefDivIgnoreBuff;
-
-/**
- * ==================================================================
- * 減算Mdef算出
- * ================================================================== 
- */
-
-	//----------------------------------------------------------------
-	// 基本減算Ｍｄｅｆ
-	//----------------------------------------------------------------
-	var mdefMinus = 0;
-	mdefMinus = Math.floor(n_A_INT + n_A_BaseLV / 4 + n_A_VIT / 5 + n_A_DEX / 5);
-
-	//----------------------------------------------------------------
-	// 特殊効果の適用（各々乗算扱い）
-	//----------------------------------------------------------------
-
-	// ウルフヘジンの時限効果
-	if (TimeItemNumSearch(TIME_ITEM_ID_WOLF_HEZIN)) {
-		mdefMinus -= Math.floor(mdefMinus * 20 / 100);
-	}
-
-	// 「アヌビス帽」の、効果（ペナルティ）
-	if (EquipNumSearch(1281)) {
-		mdefMinus = Math.floor(mdefMinus / 2);
-	}
-
-	// 「ジェネティック　Ｓホム　ペインキラー」の、効果
-	if ((bufLv = ExBuffNumSearch(EXBUF_ID_HOM_S_PAINKILLER)) > 0) {
-		var homLv = ExBuffNumSearch(EXBUF_ID_HOM_S_PAINKILLER_HOM_LEVEL);
-
-		mdefMinus += ROUNDDOWN(100 * bufLv * (132 + homLv) / 100 * n_A_BaseLV / 150);
-	}
-
-	//----------------------------------------------------------------
-	// 計算した結果をキャラクターデータに保存
-	//----------------------------------------------------------------
-	charaData[CHARA_DATA_INDEX_MDEF_MINUS] = mdefMinus;
-
-/**
- * ==================================================================
- * HIT 算出
- * ================================================================== 
- */
-	{
-		let hit = 0;
-
-		//----------------------------------------------------------------
-		// 基本ＨＩＴ
-		//----------------------------------------------------------------
-		hit = 175 + Math.floor(n_A_BaseLV + n_A_DEX + n_A_LUK / 3);
-
-		//----------------------------------------------------------------
-		// 装備、支援等による補正（＋○○）
-		//----------------------------------------------------------------
-		let hitPlus = 0;
-
-		// アイテム特性
-		hitPlus += n_tok[ITEM_SP_HIT_PLUS];
-
-		// 装備追加効果、支援効果
-		hitPlus += GetStatusModifyHitPlus();
-
-		// 最終的な効果を適用
-		hit += hitPlus;
-
-		/**
-		 * プレイヤー状態異常「暗黒」の効果
-		 */
-		if (g_confDataDebuff[CCharaConfDebuff.CONF_ID_DARKNESS] > 0) {
-			hit = Math.floor(hit * 0.75);
-		}
-		/**
-		 * プレイヤー状態異常「不幸」の効果
-		 */
-		if (g_confDataDebuff[CCharaConfDebuff.CONF_ID_UNLUCKY] > 0) {
-			hit = 0;
-		}
-
-		//----------------------------------------------------------------
-		// 計算した結果をキャラクターデータに保存
-		//----------------------------------------------------------------
-		charaData[CHARA_DATA_INDEX_HIT] = hit;
-	}
+	ApplyPlayerAtkBase(charaData, mobData);
+	ApplyPlayerMaxHpMaxSp(charaData);
+	ApplyPlayerDef(charaData);
+	ApplyPlayerMdefHit(charaData);
 
 	// FLEE の計算
 	charaData[CHARA_DATA_INDEX_FLEE] = getFlee();
